requirements.txt.txt
requests==2.31.0
pandas==2.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
apache-airflow==2.7.0
pytest==7.4.0
sqlalchemy==2.0.20